third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_webkit02_run_type=write_single-expected.txt

This is a testharness.js-based test.
[FAIL] html5lib_webkit02.html ae6f2e0a014f620269920ceb12660ec708236846
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|     <input>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <input>"
Harness: the test ran to completion.
